# Conflicts:
#	app/src/main/java/com/example/ecommerce/SignupTabFragment.java

+++ app/src/main/java/com/example/ecommerce/SignupTabFragment.java
@@ -27,7 +27,6 @@ import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
 
 import java.util.HashMap;
@@ -259,8 +258,8 @@ public class SignupTabFragment extends Fragment {
                                                 @Override
                                                 public void onComplete(@NonNull Task<Void> task) {
 
-                                                    Map< String,Object > listSize = new HashMap<>();
-                                                    listSize.put("list_size", 0);
+                                                    Map<String,Object> listSize = new HashMap<>();
+                                                    listSize.put("list_size",(long) 0);
                                                     if (task.isSuccessful()){
                                                         firebaseFirestore.collection("USERS").document(firebaseAuth.getUid()).collection("USER_DATA").document("MY_WISHLIST")
                                                                         .set(listSize).addOnCompleteListener(new OnCompleteListener<Void>() {
